Extract AssociatedRouteBuilder to own module
Primary focus here being on readability and future extensibility of the
router builder code base.
