+++ gotham/Cargo.toml
@@ -13,7 +13,7 @@ keywords = ["http", "async", "web", "framework", "blockchain"]
 
 [dependencies]
 log = "0.4.1"
-hyper = { version = "0.11.7", features = [] }
+hyper = { version = "0.11.25", features = [] }
 serde = "1"
 serde_derive = "1"
 bincode = "1"
